+++ pgdrive/scene_creator/blocks/ramp.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from pgdrive.pg_config.parameter_space import Parameter, BlockParameterSpace
 from pgdrive.pg_config.pg_space import PgSpace
-from pgdrive.scene_creator import check_lane_on_road, Decoration
+from pgdrive.utils.basic_utils import check_lane_on_road, Decoration
 from pgdrive.scene_creator.blocks.block import Block, BlockSocket
 from pgdrive.scene_creator.blocks.create_block_utils import ExtendStraightLane, CreateRoadFrom, CreateAdverseRoad, \
     sharpbend
