+++ pgdrive/scene_creator/ego_vehicle/base_vehicle.py
@@ -8,7 +8,7 @@ import numpy as np
 from panda3d.bullet import BulletVehicle, BulletBoxShape, BulletRigidBodyNode, ZUp, BulletWorld, BulletGhostNode
 from panda3d.core import Vec3, TransformState, NodePath, LQuaternionf, BitMask32, PythonCallbackObject, TextNode
 
-from pgdrive.scene_creator.basic_utils import ray_localization
+from pgdrive.utils.basic_utils import ray_localization
 from pgdrive.pg_config import PgConfig
 from pgdrive.pg_config.body_name import BodyName
 from pgdrive.pg_config.cam_mask import CamMask
@@ -426,7 +426,7 @@ class BaseVehicle(DynamicElement):
             node1 = contact.getNode1()
             name = [node0.getName(), node1.getName()]
             name.remove(BodyName.Ego_vehicle)
-            if name[0] == "Ground":
+            if name[0] == "Ground" or name[0] == BodyName.Lane:
                 continue
             elif name[0] == BodyName.Side_walk:
                 self.out_of_road = True
